[feature] Fix some bug

+++ sync.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-import settings
 import shutil
 import sys
 import inspect
@@ -29,11 +28,14 @@ def git_check_repo_exists():
     return info.split('\n')[0].split(' ')[0].split('\t')[1] == settings.HISTORY_REPO
 
 
-def git_init_and_check():
+def git_init():
     if not os.path.exists(full_path(settings.HISTORY_DIR)):
         sys_call('git clone %s %s' % (settings.HISTORY_REPO, full_path(settings.HISTORY_DIR)), showcmd=True)
     git_dir = full_path(settings.HISTORY_DIR)
     os.chdir(git_dir)
+
+
+def git_check():
     if not git_check_repo_exists():
         msg = 'repo %s not found in directory: %s' % (settings.HISTORY_REPO, settings.HISTORY_DIR)
         raise Exception(msg)
@@ -65,23 +67,27 @@ def sync():
     shells = sys.argv[1:] if len(sys.argv) > 1 else get_shells()
     Log('shells: ', shells)
 
-    git_init_and_check()
-
-    Log('==> [ pull history from origin ]')
-    sys_call('git fetch origin')
-    if not git_check_repo_bare():
-        sys_call('git reset --hard origin/master')
-
-    Log('==> [ merge histories ]')
+    git_init()
     try:
+        git_check()
+
+        Log('==> [ pull history from origin ]')
+        sys_call('git fetch origin')
+        if not git_check_repo_bare():
+            sys_call('git reset --hard origin/master')
+
+        Log('==> [ merge histories ]')
         os.makedirs(full_path(settings.BACKUP_DIR), exist_ok=True)
         if not os.path.exists(settings.HISTORY_FILE_NAME):
-            open(settings.HISTORY_FILE_NAME, 'a').close()
+            touch_file(settings.HISTORY_FILE_NAME)
 
         for shell in shells:
             Log('[ merge %s... ]' % shell)
             ShellHistory = globals()[shell.capitalize() + 'History']
             shell_his_dir = ShellHistory._get_history_dir()
+            if not os.path.exists(shell_his_dir):
+                Log('%s shell history does not exist' % shell)
+                continue
             shutil.copy(shell_his_dir,
                         os.path.join(settings.BACKUP_DIR, shell + '_history.' + get_time_now_str()))
             History.merge_file(shell_his_dir, settings.HISTORY_FILE_NAME, settings.HISTORY_FILE_NAME, cls1=ShellHistory,
@@ -93,8 +99,11 @@ def sync():
             ShellHistory.convert_file(settings.HISTORY_FILE_NAME, shell_his_dir, cls1=History)
     except Exception as e:
         LogError('[Error while syncing]')
-        sys_call('git clean -d -f')
-        sys_call('git reset --hard master')
+        try:
+            sys_call('git clean -d -f')
+            sys_call('git reset --hard master')
+        except Exception as e:
+            pass
         raise e
 
     Log('==> [ push history to origin ]')
